1. UI related changes 2. Changed amtpaid,premium and commission rate to double 3. Introduced a model for products 4. Linked products to policy page

+++ src/main/java/com/insurance/brokman/model/PaymentDetails.java
@@ -16,7 +16,7 @@ public class PaymentDetails {
     private Policy policy;
 
 
-    private long amtPaid;
+    private double amtPaid;
 
     private Date entryDate;
 
@@ -43,11 +43,11 @@ public class PaymentDetails {
 
 
 
-    public long getAmtPaid() {
+    public double getAmtPaid() {
         return amtPaid;
     }
 
-    public void setAmtPaid(long amtPaid) {
+    public void setAmtPaid(double amtPaid) {
         this.amtPaid = amtPaid;
     }
